Update patch set 1
Patch Set 1: Looks good to me (core reviewer)

so moving the order is fine, but as the screen commands is effectively a fork, this doesn't guaruntee that nova-conductor is responding by the time nova-compute needs it. Is there a mechanism for nova-compute to be patient when it comes up?

Patch-set: 1
Label: Code-Review=+2