Update patch set 1
Patch Set 1:

Steve: the fact is that we already do the "sudo pip install" trick for the requires of each openstack components. So the extras module will get installed that way.

Now, if the question is "should we really do that 'sudo pip install' thing?", that's a different topic. It's easier to not update devstack and to rely on the requires in each components, imho.

Patch-set: 1
Reviewer: Gerrit User 4460 <4460@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0