Update patch set 2
Patch Set 2:

I did the second change because I thought someone will cry If we append to the targets.conf.

But when I try to create a set by intersecting users by this conditions:
- Wants to run devstack (even with FORCE)
- Does not have conf.d by default
- Causes issues appending to the targets.conf

Probably I get an empty set.
So probably the first change set, with a little reformatting is better.

If you think, the set is not close to an empty set, I go with your recommendation.

Patch-set: 2