Update patch set 3
Patch Set 3:

Nachi: Mark and I were talking about how to best deal with this problem, and a better solution that came of our discussion was to ensure that the durability setting of all users of the notification queue is set in one place in devstack.  This ensures consistency during the initial install, and makes it harder for a single project configuration change to break other projects.

Patch-set: 3
Label: Code-Review=0