Update patch set 2
Patch Set 2:

I'm fine with this as long as we better explain the rationale in the commit message. The current explanation isn't quite right as the bug referenced in the reverted commit wasn't exactly the cause of this problem:

The attempt to set the queue durability for Glance notifications always sets the queues to durable. We are reverting this until a refined approach is available.

Patch-set: 2
Reviewer: Gerrit User 1132 <1132@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0
Label: Workflow=0