Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

(1 inline comment)

The code seems OK to me, but I have a different thought about when and how to check the service status. Feel free to ignore me.

Instead of checking the all the service status at the very end, I think maybe we should check the every single service status in the function screen_it(). As soon as it detects a service not running correctly, it should stops/exit by default and tells the users of the failure.

Besides, instead of struggling with the pid/ppid, would it be simple to change command being sent to a screen window in function screen_it() as something like:

   screen -S $SCREEN_NAME -p $1 -X stuff "rm -f $1.failure && $2 || touch $1.failure$NL"

and check the existence of the file $1.failure for failure?

Patch-set: 1
Label: Code-Review=-1