Update patch set 1
Patch Set 1:

I'm not sure I agree with Lianhao Lu about stopping devstack as soon as there's a failure. It has actually been useful to me to have devstack go on, so I could then later fix the small issue and manually start the service, without having to call ./stack.sh again.

I agree that having to deal with the pid/ppid like that is a bit painful, though. Having some other way to implement this might be good.

Oh, and outputting "Service foo is working well" is not that useful; better to just tell when there's an error, to make it the errors easier to spot.

Patch-set: 1
Label: Code-Review=0