Update patch set 2
Patch Set 2: (1 inline comment)

(replied to the inline comment)

re the exit: we actually already do that with the check for FORCE=yes. The idea behind doing exit everywhere else, with a nice message, is to make it easier to port to another distro (just run, wait for an issue, fix, run again, etc.). I think it's a bit better better than just defaulting to the behavior of one distro and assume it will work.

That being said, I don't care that much about this. I'm happy to do whatever you feel is better.

Patch-set: 2