Update patch set 2
Patch Set 2: (2 inline comments)

If we are going to specifically call out each distro in these checks then the else clause should fall through (or be absent) if it remains a valid configuration otherwise issue a useful message and stop.

This is a change from how devstack was built: mostly only testing for exceptions and defaulting to the general case.  It is also part of the reason it took a good bit of lobbying to get fedora supported in the first place as now we're well on the way to having bits of the code look like pre-autoconf portable C code.

Patch-set: 2
Label: Workflow=0