Update patch set 1
Patch Set 1:

OK, it looks the the real issue here is that Ubuntu setuptools writes to /usr/local/bin and Fedora writes to /usr/bin.  I don't know which one is "right" but my SA-gut tells me that non-system tools writing to /usr/bin is wrong; however that discussion belongs not here.

It seems to me the fix you seek is just setting *_BIN_DIR to the setuptools dest properly.  The question about still needing to support the project bin dir remains but I suspect the answer there is 'no'.

If possible I would like setting the dir variable to remain with the other defaults so it can be picked up by other projects without actually running the config function.  Currently only Nova has the potential for this being a problem in stack.sh but that is not the only script that uses the lib/* files now.

Patch-set: 1