Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

Not sure about this being a bug, but it's definitely a nice enhancement/cleanup.

The only thing I was wondering is could you go ahead and change the volumes.sh exercises as well and maybe it would make sense to move the declaration in to exerciserc?

Patch-set: 2
Label: Code-Review=-1