Update patch set 6
Patch Set 6:

Sean, I fully agree with you that a refactoring is required for the quantum related code in stack.sh. However, that is a larger effort and will hold up integrating this patch which is much smaller. I think it would be much more beneficial to the FloodLight community to have this patch available right away, than have to wait for the bigger changes to come through first. I believe this was the reason Dan W and Gary K agreed with me submitting this patch in the first place. Hence I would request you to reconsider your stand. There also seems to be support for merging this patch from some of the other reviewers.

Patch-set: 6