Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

I'm really torn about this. This adds a lot of complexity for a non core project, and definitely takes us out of the "opinionated" space.

I'm not convinced this is the direction we want to take the project, but I'll let Dean have the final word here.

Patch-set: 1
Label: Code-Review=-1
Label: Workflow=0