Update patch set 3
Patch Set 3:

It _is_ pragmatism as stack.sh was/is a mess and moving the projects out is necessary.  I had an agreement from someone on the Quantum team with +2 on devstack to do that and it is still not done.

Regarding the length of time this patch has been in-flight, if it had been ready sooner you may have missed being caught by this particular decision.

In short, devtack is not going to support every possible combination of Quantum configurations.  We need an equivalent to nova-net's FlatDHCP.  And FWIW, this sort of paring down is also going to be considered in all other areas of devstack.  It is not all things to all people, it is an opinionated development tool.

Patch-set: 3
Label: Code-Review=0
Label: Workflow=0