Update patch set 1
Patch Set 1: Looks good to me (core reviewer)

Thanks.  The use of the psuedo-service names (nova, glance, etc) is somewhat sloppy and there are likely to be more of these. This is the second in a week or so...

Patch-set: 1
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2
Label: Workflow=0