Update patch set 1
Patch Set 1:

FWIW, anytime soon would be a good time to use the SQL catalog rather than the template catalog.  SQL has been the default for a while and is what is used in the gating tests.  The template setup is likely to be removed from devstack before grizzly is released.

Patch-set: 1