Update patch set 3
Patch Set 3:

Maru, no worries.  Part of my frustration is that I've started the Quantum refactor twice but the rate of change there was enough that the rebases were eating into other things I needed to do.

Also, Xen is a special case as we (the original gang responsible for devstack anyway) have no way to test it anymore. I generally wait for one of the xen contributors to review those patches before approving them.  All I can test is that it doesn't break a non-xen config.

Patch-set: 3