Update patch set 3
Patch Set 3:

This discussion actually started during the merge request to add floodlight/bigswitch plugin support for quantum.

I'm personally in favor of using a freeze on introducing new quantum plugins to devstack to "nudge" the community to refactor quantum in devstack... in fact, Sumit (author of floodlight change) had volunteered to do this work, but wanted to merge the floodlight stuff first.  I'd prefer to see the refactoring happen first.  I will ping Sumit on this.

Patch-set: 3
Label: Code-Review=0