Update patch set 1
Patch Set 1:

afazekas: With respect, consider asking questions if something in a review doesn't make sense to you.  OpenStack is complicated, and I haven't found it useful to assume I know someone's intent unless I've recently worked on the same code.

Please note, given Sean and Dean's admonition that no further changes to stack.sh to support quantum will be accepted in advance of a refactor that externalizes such functionality, this change won't go in as-is anyway.

First issue - this change adds support for configuring a router that is publicly owned if namespaces are disabled.  Before this change, the router1 name is fixed.  This change does not change that behaviour and is not intended to - I would consider it out-of-scope.

Your second two issues are not related to this change - please respond on the review for the corresponding tempest change.   When you do, please consider the following:

- names are not guaranteed to be unique, but ids are.  when configuring tempest, the use of ids rather than names to identify resources is in general preferable to having to look up the name and potentially be returned multiple resources at test-time.

-  the default for public_router_id should be blank (as should public_network_id), but devstack's iniset will create new config entries above the blank ones, and the blank ones will win.  Adding fake values (as is used elsewhere in the sample conf) ensures that devstack can properly configure tempest.  Should iniset be fixed?  Of course.  But I'm working on the problem at hand - allow configuration of tempest - using the tools at hand using the approach already taken.  If you feel this is a problem, consider solving it at its root rather than -1'ing a patch that is entirely consistent with the existing codebase.

Patch-set: 1
Label: Code-Review=0