Update patch set 8
Patch Set 8:

I have refactored the patch to align with the quantum refactoring, stack.sh is untouched in this new patch, changes are confined to the quantum modules.

Thanks to Akihiro MOTOKI for for submitting the earlier quantum refactoring patch and making it possible to align this patch per the wishes of the reviews.

Kindly review.

Patch-set: 8
Label: Code-Review=0