Add basic uec image preparation to tempest
Add uec image preparation to lib/tempest.
cirros as image is hard coded at the moment.

If the images does not exists or the system is not able to use uec images the
image prepare step will be skipped and tempest will skip the related
tests as well.

Setting ssh username correctly.

Change-Id: I0d36ac7834e1eb677007e2c92dfc375d134a6023