Update patch set 9
Patch Set 9:

Thanks Isaku for the heads up regarding the other patch. It's been a few weeks since we submitted the original patch in this review, and have been trying to work through the reviews. If you don't mind, can we address this patch first? I can help and work with you on your patch if/when this patch gets accepted.

Patch-set: 9
Label: Code-Review=0