Update patch set 1
Patch Set 1:

Zhongyue Luo: You may want to consider reading the review history before chiming in with your opinion.  As per a previous comment, this review is unlikely to go in as-is due to a requirement on the part of devstack core that quantum changes not be accepted to stack.sh in advance of a refactor to lib/quantum.

Patch-set: 1