Add mongodb as a kind of supported database type for ceilometer
Since mongodb is used by ceilometer, this patch is to add mongodb
as a kind of database type for ceilometer convenience.

Since the other projects except ceilometer use relational databases
and support SQL only, mongodb can't be set as the database type for
them in devstack until they support it in the future.

Change-Id: I22faae2086711f1b98734d43c572dbcf1dc32adb
Signed-off-by: Hongna Xu <hongnax.xu@intel.com>
Signed-off-by: Shane Wang <shane.wang@intel.com>