Update patch set 4
Patch Set 4: I would prefer that you didn't merge this

To clarify how stackrc should be used:  it is for global configuration variables and things that need to be declared early so they can be referenced in localrc.

Project-specific configuration belongs in the project files.  In this case, that is why the inclusion of the quantum vars in tempest needs to be documented.

PUBLIC|PRIVATE_NETWORK_NAME should probably go back to lib/quantum, I don't see a reason they are global config.

@afazekas, your description is pretty close, except for the stackrc 'SHOULD" section.  Shared vars that need to be referenced before they would be sourced in their project files should go into stackrc.  We have few, if any, of those for project-specific vars.  That is more of the rule I use for moving things from stack.sh itself into stackrc.  And that is usually so they can be dereferenced in localrc, like $DEST.

Patch-set: 4
Label: Code-Review=-1
Label: Workflow=0