Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

(1 inline comment)

To repeat a bit from the inline comment, the $STACK_USER value should only be used as the target user when running as root.  Now that this exists we need to watch out for code that wants to use it in place of the current user, which may not be stack.  I don't consider starting stack.sh as root and switching to $STACK_USER to be the 'normal' case so we can't make that assumption elsewhere.

Patch-set: 2
Label: Code-Review=-1
Label: Workflow=0