Update patch set 4
Patch Set 4:

@dtroyer:

I recommended to put the project specific variables into the stackrc because you MAY touch the same variables in the localrc.

I assume from your other review response, generally we don't want environment variables, to take precedence, because you asked me for direct declararation.

I think variables expected optionally declared in localrc, should be enumerated somewhere. Probably the stackrc is the best place for that. And it is the best place to set the default value IMHO.

When you are searching for where is XYZ variable gets a value before on optional {:-} assignment, you find nothing otherwise.

Difficult to figure out is the {:-} done by a mistake, or a localrc declaration expected.

Anyway I think this change should pass, and we should discuss the variable declaration and cleanups on IRC.

I will help in making it consistent anyway, we just need to agree on something.

Patch-set: 4