Update patch set 1
Patch Set 1:

Interesting stuff...

Please structure this similarly to the other projects in lib/, it should be in its own file (lib/ldap).  The .ldif files should go in files/ or files/ldap/ if you prefer.

I saw a comment regarding making it a part of keystone; I'd prefer to not do that as it looks like you are using the packaged openldap server so it is a system service.  If configured to use ldap keystone should test for it being available but I'd think that is all.

Patch-set: 1
Label: Code-Review=0
Label: Workflow=0