Update patch set 1
Patch Set 1:

The file master is never created within that directory. If you look at the code, the variable TOP_DIR is pointing to the directory, in which install_os_domU.sh is in. If you search for a code for master, you won't see any occurence of master, apart from the one you spotted. So, to answer your question, I don't know what is that master. My guess is, that if you look at the documentation, a file called master contains the devstack zipball. However, that's out of the scope of this script. I think the master there is just a bug. The change that introduced this is:
https://review.openstack.org/#/c/6317/8
I would go for yagni

Patch-set: 1