Update patch set 2
Patch Set 2:

Sean,  Yes, it is intentional that LDAP triggers the Keystone config, and still requires the Database.  You cannot run Keystone without a Database, even if you use LDAP for the Identity store, as LDAP does not store Policy, Service Catalog nor the Tokens.

LDAP should be an optional component, and not installed by default, but if it is is indicated that it is enabled, the above logic should trigger it.  It is much like postgres or qpid;  they are not installed by default, but used if they are indicated on the command line.

Patch-set: 2
Label: Code-Review=0