Update patch set 2
Patch Set 2:

right, though it is a little different. In the qpid or pg case they are implementations of a feature (database or queue) that you can't function without. ldap is a keystone backend implementation detail, like the quantum plugins.

So it seems like KEYSTONE_BACKEND=ldap (or something) should be the stackrc and that triggers this to run.

Patch-set: 2