Fix for multiple compute nodes.
If an attempt is made to start a secondary node, such as a
secondary compute node, the current version will incorrectly
attempt to setup a database and rpc backend on the secondary
node.  Setting up a secondary database is catastrophic as it
will cause the nova database setup by the primary to be dropped.
Attempting to setup a secondary rpc backend is incorrect because
the primary node has previously set one up.

Change-Id: I494c17b7ab8fc85d8704921492036ea10a41e00e