Update patch set 6
Patch Set 6: Looks good to me, but someone else must approve

I think the only thing that's missing is documentation on the 2 variables you need to enable to make this work. I'm going to +1 for now, +2 once we get something in here to let people know how they would enable the feature.

Patch-set: 6