Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

I think I would rather put the individual swift services in ENABLED_SERVICES like all of the other projects and add swift to the list in is_service_enabled().  Making swift be even more different in service handling does not seem like an improvement to me.

There would be a compatibility issue that can be handled in stack.sh with a substitution of the services for the string 'swift' in ENABLED_SERVICES.

Patch-set: 3
Label: Code-Review=-1
Label: Workflow=0