Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

Can we default to not including the option on the dnsmasq command line if a valid value is not set?  I don't like the idea of defaulting to sending packets to an outside DNS server that some organizations explicitly block. Or grabbing a server from /etc/resolv.conf?

Patch-set: 2
Label: Code-Review=-1
Label: Workflow=0