Update patch set 2
Patch Set 2:

Sean, 

I am happy to split the patches, as it makes sense.
I have to be honest I did not think about the right syntax. I just picked the first which worked.

Are you referring just to the syntax of the or clause, or do you think we should work out better the logic of ini_set when setting the same option multiple times?

Patch-set: 2