Update patch set 1
Patch Set 1:

Dean,

I had the same thought but as the new function would share most of the logic of iniset, then I thought it migth be worth updating it.

I understand your concern and I'm not going to keep pushing it.
I don't understand if you're happy with a iniadd, or if you want this bug to be marked as invalid and find other solutions for code which needs to set a value multiple times.

Patch-set: 1