Update patch set 1
Patch Set 1:

Hi Dean,

as much as I can agree with you that multi string and list options are at least redundant, We have one of those options in the nicira quantum plugin. Since I think here and there these options are used, I initially thought that it might be good to do a change in devstack's infrastructure instead of a hack in the nicira quantum plugin's configuration code.

On the other side, I'm really not willing to write full support for handling multi string options in ini files. Treating them as arrays is going to be far from easy, and is definitely error prone.

So if nobody shows up in 24 hours and says they'd support multi string options too, I'll abandon this patch, mark the bug as invalid, and push a different solution for the quantum nicira plugin.

Patch-set: 1