Update patch set 2
Patch Set 2:

It's the subtleties of the database mechanism...it may bee too clever for its own good and it's going to be a problem in Grenade.  I'm going to take a look at simplifying it so just putting a supported database name in ENABLED_SERVICES is enough.

Patch-set: 2
Label: Code-Review=0
Label: Workflow=0