Update boot_from_volume.sh
This change updates the BFV exercise to use the new'ish
cinder create with image-id.  This is a pretty simplified
version from now, and simply ensures we create a bootable
volume, and that nova can boot an instance on it.

We're not doing an networking checks, although we can look
at adding those before removing the skip on this in the gate tests.

Change-Id: I37bff49e0bde14b5f7bb428f2b0fc5a5680c8426