Add option to make screen starting more robust.
We have seen a number of failures in ci where a host is overloaded
and the 1.5 second sleep before stuffing data into screen is not
long enough. This means the service doesn't start and tests fail.

This change adds a config option to allow us to turn off the developer
friendly option to stuff text into the screen. When SCREEN_DEV is
set to False it will use a simple exec in screen instead of stuff.
This should be far more reliable because we don't have to wait
for bash to start.

Change-Id: I7f1b5dbf5329b23507cb767d54a2795be0d73e01