Update patch set 4
Patch Set 4:

In a multi-node configuration would there be a different set of swift services started on the non-master nodes?  That is one of the primary reasons the other project services are listed individually. I understand about the compatibility and can probably live with it.

Patch-set: 4
Label: Code-Review=0
Label: Workflow=0