Update patch set 1
Patch Set 1:

so I dug into the problem with ComputeCapabilitiesFilter a bit more, and the problem is that we're specifying deploy_[kernel|ramdisk]_id on the instance_type extra_specs, but we're *not* advertising the same values within the baremetal driver.

The CCB is there to ensure that extra_specs:cpu_arch is matched properly, and it does this quite well (we probably don't need that check inside the hypervisor driver, actually).

The solution to this apparent bug is to amend the [baremetal] instance_type_extra_specs config option with the uuid's of the deploy images. However, in devstack, nova must have started before we generate those uuid's... chicken-vs-egg...

Patch-set: 1
Label: Code-Review=0