Update patch set 3
Patch Set 3:

My only problem with it is the removal of any sort of test to verify that the instance has in fact booted.  I agree that Tempest should be the focus of any testing more than the sanity checks that DevStack does, but not at the expense of the DevStack exercises in the process.

Patch-set: 3
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0
Label: Workflow=0