Update patch set 4
Patch Set 4: Looks good to me (core reviewer); Approved

Indeed, was waiting for Jenkins.

To document what Stephen and I chatted about on IRC, we'll let the user configure c-bak without swift as other backends external to DevStack are supported.  Attempts to use Swift in that care are reported to the user with an error message that indicates the lack of Swift.

Patch-set: 4