Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

It looks good to me. I only have a pedantic comment about

[ "$CINDER_MULTI_LVM_BACKEND" == "True" ]

which should be 

[ "$CINDER_MULTI_LVM_BACKEND" = "True" ]

to be consistent with the other [ ] expressions. I did not know that [ ] allowed ==, I thought it was only allowed in the [[ ]] shell builtin ;-)

Patch-set: 3
Label: Code-Review=+1