Add option to launch services w/out screen.
This change is to test an idea of launching services in
devstack without using screen.

As an initial try, we try to do this with as minimal impact
and change as possible.  If we determine that this is useful and
we want to go this route long-term there's definitely some optimizing
and clean up that can and should be done.

As it stands right now, this patch just looks for a "NO_SCREEN=True"
variable to be set and if it's set to True, then rather than launch
in a screen session processes are launched using "bash -c".

Logs should be stored in the same format and in the same location as
the symlinks for the current gate runs.

Capturing pids in this case is a bit tedious, so for unstack we just
gather up pgrep resutls for each installed bin service
(ie "pgrep -f /opt/stack/*/bin) and kill all of the pids we find
in the results.

Change-Id: I19f84a248d1a636a8c3a15de85c1280bb7d2af5d