Update patch set 1
Patch Set 1: (8 inline comments)

There is a bootstrapping problem with having multiple files - there are too many different localrc's in XenServer installation guides, in part because the file has to exist in parallel with the script that's performing the installation.

I'm keen to have a single file that can be used _outside_ of dom0 to setup everything in dom0 and devstack.

This will mean that the installation instructions are "wget file XYZ and run it with ABC arguments" rather than download the whole of devstack then cd to devstack/*/tools/xen then create a localrc - only for devstack to be cloned to where it's really needed, in the DomU.

I don't think it's as desirable but on reflection of the comments it seems that devstack isn't the appropriate place for this script - so if we still want to, we can look at hosting it elsewhere

Patch-set: 1
Reviewer: Gerrit User 6735 <6735@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0