use paste instead of stuff
the screen man page seems to say stuff is not suitable for long
strings (it is instead "performant"). Instead of using stuff we
can use the register / paste combo. This may have an effect on races
that we are seeing.

Paste also will let us control the paste rate, if we believe that
would help with the non screen start up behavior.

Change-Id: Ib5ddc829e91bb630b1d06d36f9817234194ef31a