Add some checks for screen creation and activity.
Cases of both screen session never starting,
and empty screen logs have been noted for a number
of services in the gate tests.

This patch adds a simple check on screen_it to make
sure that the screen session was launched by checking
for the SCREEN_LOGDIR file.

As a last step in stack.sh we also check the screen logs
and make sure they're not empty.  This doesn't fix
devstack bug 1133355, but it will help us detec it sooner
and avoid erroneous data collection due to assigning recheck
bugs to what appears to be the problem but isn't root cause.

Change-Id: I1193aea501a7babe27bdb4119fc85814bd04c0d0
NOTE:  This change ONLY works when using SCREEN_LOG_DIR setting