Update patch set 6
Patch Set 6:

As I stair at this code, I'm wondering if we are going after the fix in the wrong place.

First, I don't actually understand why we are symlinking at all. Anyone have history on that? Also, the symlink can't fail, it's a -f, so even if the source file isn't there, the symlink is going to exist.

I feel like it's much more likely that screen is failing to handle the "-X log on" because of load. That would explain a lot.

But lastly, why are we using screen logs as our capture point? Wouldn't it be better to actually set up file logging for these daemons (which would be more like one might expect on a real system?)

Patch-set: 6
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0
Label: Workflow=0