Update patch set 1
Patch Set 1: Work In Progress

This should be functional, it runs in my VM.  The unbuffering of Python's stdout bothers me a bit yet, but I don't want to hold things up over that unless it turns out to be a big issue.

Also, what is the feeling about renaming all of the 'screen' stuff like log files, directories, etc?  Do it now or later?  I've wanted to get screen out of the names for a while now...

Patch-set: 1