Allow DATABASE_TYPE to be specified in localrc
The use of DATABASE_TYPE without forcing it into ENABLED_SERVICES
seems to be the only way to set up a multi-host devstack installation.

If there are other ways to set up a multi-host installation, then this
change may not be needed.

Also, we should look at whether DATABASE_TYPE should be deprecated or
if we can safely just deprecate use_database

Change-Id: I4c6e8ee31f298004338ca46bfc330a3e3af288aa
Fixes: bug 1136028