Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

(1 inline comment)

I think what Sean is saying is that in addition to the use case you've implemented where the caller knows all of the values, it would be useful to have a way to add a new value to a multi-line option without the caller having to manage all of the old values, too. An iniadd could be implemented by calling iniget_multiline to find the current values, then using iniset_multiline to update the settings with the new value at the end of the argument list.

Patch-set: 1
Label: Code-Review=-1