Allow mongod service to launch on Fedora
Fixes bug 1144907

Previously the mongod service failed to launch, as the eager
filesystem freespace check tripped over an SELinux AVC and
also possibly a lack of freespace (depending on the size of
the hosting instance).

We work around these issue by ensuring SELinux is in
permissive mode on Fedora and also setting the --smallfiles
option.

Change-Id: I74f21662aedf9cc07b9a335d22b833a8a80dfce2