Update patch set 2
Patch Set 2:

It turns out that the required policies are provided in the latest version of the selinux-policy-targeted package.

So we can address the issue in a different way by ensuring that the selinux-policy-targeted RPM is up-to-date when ceilometer is enabled.

So this should address the concerns expressed about the bluntness of making SELinux permissive.

Cheers,
Eoghan

Patch-set: 2
Label: Code-Review=0