Update patch set 12
Patch Set 12: Looks good to me, but someone else must approve

I guess it is better that ability of a plugin is indicated with a flag such as Q_PLUGIN_HAS_SG instead of a function. So the plugin that is unrelated to the feature can completely ignore it.

Any way, this patch works fine.

Patch-set: 12
Reviewer: Gerrit User 4726 <4726@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1