Force Quantum processes to run under bash
Force bash to evaluate the command line passed to screen_it so the
commands are executed in a subshell.  This seems to make a difference
when using run_service() when the server process is owned by init.

These Quantum services were the only ones in all of DevStack being
spawned in this manner.  This will allow the run_service() patch
https://review.openstack.org/#/c/23148/ to succesfully be merged
and the number of test failures due to screen should be reduced to 0.

Change-Id: I97d562adda51d39a5e7fa1dc4d945d4a396201cd