Add is_rhel6
Add is_rhel6 function to test for RHEL6 as the first step for getting
support.  RHEL6 is considered to be a special version of Fedora, so
is_fedora still matches on RHEL6.  This avoids a lot of code churn,
but does mean you have to be slightly careful with ordering of checks
to get RHEL6 specific behaviour.

Change-Id: Ida19130ce9499d4d1fd2360fadc6b870b24bbc95