Update patch set 1
Patch Set 1: Do not merge

This makes a change in the assumptions about the REPO and BRANCH variables in that if they happen to be set when stackrc is sourced, their values will not be reset as they are today.  I don't know yet if this is a problem or not but it needs to be checked out.  My primary worry right now is in Grenade, although there are times when stackrc is sourced more than once in DevStack.

I think it will be likely that the only time OS_BRANCH will have the same value for all repos is for 'master'.  Specifically, the clients do not have 'milestone/proposed' or 'stable/grizzly' or whatever.

Also, thus far DevStack does not use any variables prefixed 'OS_', those are generally expected to be consumed by the command-line clients.

Patch-set: 1
Label: Code-Review=-2