Update patch set 1
Patch Set 1: (2 inline comments)

IMHO the primary change needed here is to add the Red Hat/CentOS test to GetDistro().  If the name of is_fedora() needs to change to better reflect that it is a family test, then let's consider that.  Otherwise just use $os_VENDOR and $os_RELEASE to test for specific releases.

Patch-set: 1