Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

(1 inline comment)

Why check for the home directory alone?  Why look for 0700 specifically?  If the goal is to ensure that the destination can be used by httpd, it would probably make more sense to check the permissions in a generic way, since there is nothing stopping a user from choosing a path with different but still unusable permissions that is not under /home.

Patch-set: 3