Update patch set 5
Patch Set 5:

The intent is to satisfy requests from people who have rhel as their basic standard environment for an easy path to using top-of-tree upstream for both testing and development.

For the most part the change is a specialization of the existing Fedora path; although I did create a separate rpm-list to avoid the existing one becoming too tangled up. 

I wanted to get the review out for comments early; internally I'm working on some continuous-integration testing of devstack/rhel so that we detect ongoing issues.

Patch-set: 5