Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

So it's not that I don't want to make it easy to test drivers with devstack, I just don't like doing it this way.  I think it can get out of control rather quickly in terms of lib/cinder getting very messy.

I think we can do this in a more generic fashion, where localrc variable is used for *whatever* driver instead of hard-coding all of them in the cinder file.

I'll let the devstack pros determine if I'm off base here or not, but that's my opinion.

Patch-set: 1