Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

It doesn't actually hurt anything to put the Require line in everyone's config instead of building new custom logic for fedora 18.

My preference is to not and the code, and just update everyone's config with that stanza, as it's less debt for us to carry in the code.

Patch-set: 1
Label: Code-Review=-1