Removes "RPC not enabled" error message when RPC backend is not needed
When no service needing a RPC backend is activated, no error message should
appear if a RPC backend is not installed. A simple check is done on the
services installation files to see which services need to initialize a RPC
backend at some point; if none of these services are in ENABLED_SERVICES
then the error message is skipped.

Change-Id: I4e47e0c675c74775b4ea53a00848ac1d777f0125
Fixes: bug #1167338