Update patch set 1
Patch Set 1:

Ian, thanks for breaking these back up, but they are still dependent on each other.  I'm guessing these are in the same branch so git-review creates the dependent reviews.  This makes it hard to review and merge the simple things until everything in front of them is merged.  For changes that really do not depend on each other if you can commit them from different branches they will not be marked as dependant in Gerrit.

Patch-set: 1
Label: Code-Review=0