Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

git-review on precise is 1.12. Honestly, I don't think that installing a distro tool that old is going to be incredibly useful to folks, and means another thing that has to be added to the gate jobs.

Patch-set: 1
Label: Code-Review=-1
Label: Workflow=0