Update patch set 2
Patch Set 2: Do not merge

A while back some folks did a good bit of work to make DevStack usable with selinux on Fedora.  Has that bitrotted and doesn't work any more?

I'm not against disabling it but if it is just a matter of fixing one specific thing I'd rather fix it than throw away the previous work.

Of course, if we fix it we'll need someone to champion keeping it working.  It isn't something we test regularly.  Obviously.  So did I just talk myself into a corner?

Patch-set: 2
Label: Code-Review=-2
Label: Workflow=0