Update patch set 1
Patch Set 1:

It might not be the same. For example, we are using an squid-deb-proxy to speed up the Ubuntu install, but that proxy is limited to some websites. It really depends on the user. Having these as separated parameters gives more flexibility.

Patch-set: 1
Reviewer: Gerrit User 5044 <5044@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0